make available bytes per pixel calculations that return nonintegral values
git-svn-id: http://code.imagej.net/svn/imagej/trunk@1839 be23795f-a60e-40d4-ba19-055798872e57

+++ _ij1-patches/src/main/java/ij/process/ImageProcessor.java
@@ -2254,7 +2254,7 @@ public abstract class ImageProcessor extends Object {
 	// NEW METHODS FOR IJ 2.0 SUPPORT
 	
 	public abstract int getBitDepth();
-	public abstract int getBytesPerPixel();
+	public abstract double getBytesPerPixel();
 	public abstract ImageStatistics getStatistics(int mOptions, Calibration cal);
 	public abstract boolean isFloatingType();
 	public abstract boolean isUnsignedType();
